Allow cancelling of a waitgroup
The following allows the cancelling of a WaitGroup. When a publish
call is called, it enqueues the information onto a subscriber. This
includes the WaitGroup.Done method. If the Done method is never called,
the done channel is never closed and then goroutine is never exited.
This manifests itself as leaking gorountines.

The code changes are as follows, allow the passing in of a context, that
context can be a cancellable or a timeout context. When the context done
channel is invoked we return out and exit out of the goroutine. The
result channel was changed from a struct to an error, so that it's
possible to be notified if the context was cancelled.

This should prevent the leaking of the goroutines. The subscriber side
should then ensure that they also handle cancellations.