Return WaitGroup Wait function instead of channel.
Most if not all call sites do not wait for Publish to return, they treat
the call as a fire and forget. Unfortunately the code in place always
ensures that the WaitGroup is done and returns a channel. Placing a
large burden on the callback handlers to behave. This has been proven
not to be the case and we can hold on to the goroutine. This looks like
a leak from the graphs, but in reality handlers are just not completing
in a timely manor and we're left with back pressure on the subscribers.