Metrics: Expose underlying queue status
The following exposes the underlying status of a subscriber queue. The
queue per subscriber can be susceptible to back pressure from the slow
callbacks. We currently don't have any visibility of how this manifests
as we don't have any metrics to observe. The following commit exposes
this in convenient methods that each hub can implement.

Implementations of the metrics interface should be careful around data
races, as the metrics type walks across many goroutines.