Metrics: Expose function name
Rather than exposing the index of the subscriber, expose the name of the
callback. We can do this using the runtime.FuncForPC. Although not
ideal, we get the function name at subscribe time and not at publish
time, so shouldn't impact performance during publishing.

The getFunctionName attempts to return a stable function name depending
on the type of callback passed in. Functions, methods and high order
functions should be covered in tests.

Additionally unsubscribe just uses an anonymous function to simplify the
code, as we no longer require to create unsubscriber handlers for every
subscribe.